btc: add hex support to sign message. Move hex utils from ethereumUtils to formatUtils.
