+++ src/js/core/methods/EthereumGetAddress.js
@@ -3,8 +3,9 @@
 import AbstractMethod from './AbstractMethod';
 import { validateParams, getFirmwareRange } from './helpers/paramsValidator';
 import { validatePath, getSerializedPath } from '../../utils/pathUtils';
-import { getNetworkLabel, stripHexPrefix } from '../../utils/ethereumUtils';
+import { getNetworkLabel } from '../../utils/ethereumUtils';
 import { getEthereumNetwork, getUniqueNetworks } from '../../data/CoinInfo';
+import { stripHexPrefix } from '../../utils/formatUtils';
 
 import * as UI from '../../constants/ui';
 import { UiMessage } from '../../message/builder';
